requirements.txt
flaat~=0.14.5
flask-login~=0.5.0
flask-smorest~=0.34.0
